requirements.txt
# Core pipeline: pattern → extraction → construct → inference (see README).
# For GPU PyTorch, pick a build from https://pytorch.org

python-dotenv>=1.0.0
aiohttp>=3.9.0
chromadb>=0.5.0
networkx>=3.3
numpy>=1.26.0,<2.0
pandas>=2.2.0
pillow>=10.4.0
requests>=2.32.0
scipy>=1.13.0
tqdm>=4.66.0
transformers>=4.45.0
torch>=2.2.0
torchvision>=0.17.0
google-generativeai>=0.8.0
